Move type into its own file
